### 2019-11-11 #1883 @whedon
[PRE REVIEW]: Audiometry: A Model-View-ViewModel (MVVM) Application Framework for Hearing Impairment Diagnosis
**Submitting author:** @drwaseemsheikh (<a href="http://orcid.org/0000-0002-4647-4565">Waseem Sheikh</a>)
**Repository:** <a href="https://github.com/drwaseemsheikh/audiometry.git" target ="_blank">https://github.com/drwaseemsheikh/audiometry.git</a>
**Version:** 1.0.0
