### 2019-11-11 #1883 @kyleniemeyer
Hello @drwaseemsheikh, thanks for your interest in JOSS.

Before we get started, I took a quick look at your article and software, and while it looks very useful, I can't tell from either how it satisfies one of our requirements to be in scope:
> The software should have an **obvious** research application.

Can you clarify this? Thanks!
