### 2019-11-11 #1883 @drwaseemsheikh
Thank you for your feedback.

I would like to clarify this based on the scope statement of the journal reproduced below:
“JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar.”

First, I would like to make the case that “Audiometry solves a complex problem in the domain of medicine, in particular, ear-related pathologies”. It generates important metrics including speech perception scores, hearing impairment scores, hearing disability scores, speech recognition threshold, speech discrimination score, and speech discrimination intensity, from the raw hearing test data. These scores help in better diagnosis of a particular kind of hearing loss.

Audiometry also generates complex visualization plots including pure-tone audiogram, speech audiogram, tympanogram, and calorigram based on the raw hearing test data, which are essential for an otolaryngologist to diagnose various ear-related pathologies. The shapes and relative positioning of these curves help in the diagnosis of various ear-related pathologies.

Secondly, I would like to make the case that Audiometry “supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets”.
Audiometry provides a comprehensive hearing test data management and diagnosis system to support the functioning and execution of various hearing test-related hardware. An otolaryngologist uses a wide variety of specialized equipment to gather hearing test data. However, there is no unified and centralized open-source software which provides an ability to store this vast array of hearing test data in one application. By having this data in one application, allows the otolaryngologist to more accurately and swiftly diagnose a hearing loss.

The knowledge extraction parts come through the generation of important hearing loss metrics including speech perception scores, hearing impairment scores, hearing disability scores, speech recognition threshold, speech discrimination score, and speech discrimination intensity and the generation of complex visualization plots including pure-tone audiogram, speech audiogram, tympanogram, and calorigram.
