+++ paper/paper.bib
@@ -6,7 +6,7 @@
 
 @misc{SummaryHealthStatistics:2017,
   title={Summary health statistics: national health interview survey, 2017},
-  url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/NHIS/SHS/2017_SHS_Table_A-6.pdf},
+  url={https://ftp.cdc.gov/pub/Health_Statistics/NCHS/NHIS/SHS/2017_SHS_Table_A-6.pdf},
   note={Accessed: 2019-08-07}
 }
 
