### 2019-11-13 #1883 @labarba
@drwaseemsheikh — Thank you for your clarifications. You have made the case for the software's use in clinical settings. Can you describe its use in *research* situations?
