### 2019-11-13 #1883 @drwaseemsheikh
Thank you for giving us a chance to explain some of the research applications of Audiometry.

Following are some of the research questions that can be investigated by the use of Audiometry. The coauthor, who is an otolaryngologist, is using the software in some of the following research studies:

1.	Audiometry can be used to compare the sensitivity (true-positive rate) and specificity (true-negative rate) of various hearing test equipment and methods. For example, questions like how reliable is a pure-tone audiometry test performed by a smartphone or a tablet when compared with a benchmark calibrated audiometer can be easily answered by using the proposed software.

2.	The software can be used to determine the correlation between certain lifestyle and/or work conditions and the types of hearing loss. For example, the coauthor is using the software to record hearing test data from air force servicemen who are subject to very loud noise. Using the software important correlations can be determined for example, what kind of hearing loss is common among air force servicemen. In addition, the coauthor is also investigating certain patterns and correlations among the demographics of young children in remote areas in underdeveloped countries and the type of hearing loss.

3.	The software can be used to measure the efficacy of a certain treatment, intervention, or equipment on the progression of a hearing loss.
