### 2019-11-13 308dd35
Added some research applications of the software.
