+++ paper/paper.md
@@ -30,6 +30,7 @@ Around 466 million people worldwide (over 5% of the world's population) have dis
 ``Audiometry`` can store, search, analyze, print, and visualize data corresponding to tuning fork tests including Weber, Rinne, Schwabach, absolute bone conduction, Teal, and Gelle; speech audiometry; pure-tone audiometry (PTA); impedance audiometry; bithermal caloric test; and advanced tests including alternate binaural loudness balance (ABLB), short increment sensitivity index (SISI), tone decay, and Stenger [@KramerAudiology:2019; @GelfandEssentials:2016; @KatzHandbook:2015; @DhingraDiseases:2014; @BessAudiology:2003]. The application framework can also be used to develop new hearing test applications by extending its current functionality. ``Audiometry`` is independent of specific hearing test hardware thus making it possible to be used with a wide variety of hearing test hardware. In addition, ``Audiometry`` provides a unified and uniform interface for storing, analyzing, and visualizing data from a wide range of hearing tests which traditionally rely on different hardware and software. The software was evaluated by an otolaryngologist who found it to be very beneficial in reaching a hearing impairment diagnosis conclusion more methodically, swiftly, and accurately.
 
 Following are examples of some of the research questions that can be investigated by the use of Audiometry:
+
 1.	``Audiometry`` can be used to compare the sensitivity (true-positive rate) and specificity (true-negative rate) of various hearing test equipment and methods. For example, questions like how reliable is a pure-tone audiometry test performed by a smartphone or a tablet when compared to a benchmark calibrated audiometer can be easily answered by using this software.
 
 2.	The software can be used to determine important correlations between lifestyle, work conditions, and demographics; and the types of hearing loss.
