### 2019-11-17 bd2e577
Added department in affiliation
