### 2019-11-17 f44b9fe
Update paper.md
