### 2020-01-09 #1883 @waseemsheikh
Hi,
I was just checking the status of my paper. It's been a while and it's still not being assigned to a reviewer. Thank you for providing me an update on its status.
