### 2020-01-16 drwaseemsheikh/audiometry#1 @dvberkel
whedon is reporting on missing references
@drwaseemsheikh I am reviewing your article [Audiometry: A Model-View-ViewModel (MVVM) Application Framework for Hearing Impairment Diagnosis](https://github.com/openjournals/joss-reviews/issues/2016#)

and whedon is reporting on [missing references](https://github.com/openjournals/joss-reviews/issues/2016#issuecomment-574915935). Could you take a look if whedon is missing something, or that it is correctly reported on the missing references.

Thanks.
