### 2020-01-17 drwaseemsheikh/audiometry#1 @waseemsheikh
The references were compiled using BIBTEX and I gathered all the necessary documentation for each paper using Google Scholar BIBTEX entry and sometimes there are some fields missing in the BIBTEX entry but I don't think that those fields are critical for citation. At least, I don't know how to get that information besides using Google Scholar BIBTEX citation and I followed that.
