### 2020-01-17 drwaseemsheikh/audiometry#1 @waseemsheikh
It's complaining on the missing DOIs but not all papers have their DOIs available. At least, I couldn't find all DOIs for the references. However, all the other information is in the correct BIBTEX format and complete.

Waseem
