### 2020-01-17 drwaseemsheikh/audiometry#1 @drwaseemsheikh
The references were compiled using BIBTEX and I gathered all the necessary documentation for each paper using Google Scholar BIBTEX entry and sometimes there are some fields missing in the BIBTEX entry but I don't think that those fields are critical for citation. At least, I don't know how to get that information besides using Google Scholar BIBTEX citation and I followed that.

It's complaining on the missing DOIs but not all papers have their DOIs available. At least, I couldn't find all DOIs for the references. However, all the other information is in the correct BIBTEX format and complete.

Waseem
