### 2020-01-22 4a0252b
Moved the Docs folder to the parent folder and renamed it to docs.
