+++ Docs/Doxygen/latex/a00380.tex
@@ -1,16 +1,0 @@
-\hypertarget{a00380}{}\doxysection{Audiometry/\+View\+Model/\+Search\+V\+M/\+Search\+Window\+VM.cs File Reference}
-\label{a00380}\index{Audiometry/ViewModel/SearchVM/SearchWindowVM.cs@{Audiometry/ViewModel/SearchVM/SearchWindowVM.cs}}
-\doxysubsection*{Classes}
-\begin{DoxyCompactItemize}
-\item 
-class \mbox{\hyperlink{a00984}{Audiometry.\+View\+Model.\+Search\+V\+M.\+Search\+Window\+VM}}
-\end{DoxyCompactItemize}
-\doxysubsection*{Namespaces}
-\begin{DoxyCompactItemize}
-\item 
-namespace \mbox{\hyperlink{a00401}{Audiometry}}
-\item 
-namespace \mbox{\hyperlink{a00415}{Audiometry.\+View\+Model}}
-\item 
-namespace \mbox{\hyperlink{a00426}{Audiometry.\+View\+Model.\+Search\+VM}}
-\end{DoxyCompactItemize}
