+++ Docs/Doxygen/latex/a00149.tex
@@ -1,14 +1,0 @@
-\hypertarget{a00149}{}\doxysection{Audiometry/obj/x86/\+Debug/\+View/\+Registration\+Window.g.\+cs File Reference}
-\label{a00149}\index{Audiometry/obj/x86/Debug/View/RegistrationWindow.g.cs@{Audiometry/obj/x86/Debug/View/RegistrationWindow.g.cs}}
-\doxysubsection*{Classes}
-\begin{DoxyCompactItemize}
-\item 
-class \mbox{\hyperlink{a00648}{Audiometry\+::\+View\+::\+Registration\+Window}}
-\begin{DoxyCompactList}\small\item\em \mbox{\hyperlink{a00648}{Registration\+Window}} \end{DoxyCompactList}\end{DoxyCompactItemize}
-\doxysubsection*{Namespaces}
-\begin{DoxyCompactItemize}
-\item 
-namespace \mbox{\hyperlink{a00401}{Audiometry}}
-\item 
- \mbox{\hyperlink{a00410}{Audiometry\+::\+View}}
-\end{DoxyCompactItemize}
