+++ Docs/Doxygen/latex/a00936.tex
@@ -1,42 +1,0 @@
-\hypertarget{a00936}{}\doxysection{Audiometry.\+View\+Model.\+Pure\+Tone\+V\+M.\+Pure\+Tone\+A\+M\+V\+M.\+Loud\+Discom\+Lvl\+VM Class Reference}
-\label{a00936}\index{Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM@{Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM}}
-
-
-Inheritance diagram for Audiometry.\+View\+Model.\+Pure\+Tone\+V\+M.\+Pure\+Tone\+A\+M\+V\+M.\+Loud\+Discom\+Lvl\+VM\+:
-\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[height=550pt]{a00935}
-\end{center}
-\end{figure}
-
-
-Collaboration diagram for Audiometry.\+View\+Model.\+Pure\+Tone\+V\+M.\+Pure\+Tone\+A\+M\+V\+M.\+Loud\+Discom\+Lvl\+VM\+:
-\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[height=550pt]{a00934}
-\end{center}
-\end{figure}
-\doxysubsection*{Public Member Functions}
-\begin{DoxyCompactItemize}
-\item 
-\mbox{\hyperlink{a00936_a450d8c51269778696bb09620c356fa34}{Loud\+Discom\+Lvl\+VM}} ()
-\end{DoxyCompactItemize}
-\doxysubsection*{Additional Inherited Members}
-
-
-\doxysubsection{Constructor \& Destructor Documentation}
-\mbox{\Hypertarget{a00936_a450d8c51269778696bb09620c356fa34}\label{a00936_a450d8c51269778696bb09620c356fa34}} 
-\index{Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM@{Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM}!LoudDiscomLvlVM@{LoudDiscomLvlVM}}
-\index{LoudDiscomLvlVM@{LoudDiscomLvlVM}!Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM@{Audiometry.ViewModel.PureToneVM.PureToneAMVM.LoudDiscomLvlVM}}
-\doxysubsubsection{\texorpdfstring{LoudDiscomLvlVM()}{LoudDiscomLvlVM()}}
-{\footnotesize\ttfamily Audiometry.\+View\+Model.\+Pure\+Tone\+V\+M.\+Pure\+Tone\+A\+M\+V\+M.\+Loud\+Discom\+Lvl\+V\+M.\+Loud\+Discom\+Lvl\+VM (\begin{DoxyParamCaption}{ }\end{DoxyParamCaption})\hspace{0.3cm}{\ttfamily [inline]}}
-
-
-
-The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
-\item 
-Audiometry/\+View\+Model/\+Pure\+Tone\+V\+M/\+Pure\+Tone\+A\+M\+V\+M/\mbox{\hyperlink{a00347}{Loud\+Discom\+Lvl\+V\+M.\+cs}}\end{DoxyCompactItemize}
