+++ Docs/Doxygen/latex/a00402.tex
@@ -1,31 +1,0 @@
-\hypertarget{a00402}{}\doxysection{Audiometry.\+Converter Namespace Reference}
-\label{a00402}\index{Audiometry.Converter@{Audiometry.Converter}}
-\doxysubsection*{Classes}
-\begin{DoxyCompactItemize}
-\item 
-class \mbox{\hyperlink{a00468}{Boolean\+And\+Converter}}
-\item 
-class \mbox{\hyperlink{a00472}{Boolean\+Or\+Converter}}
-\item 
-class \mbox{\hyperlink{a00476}{Enum\+Boolean\+Converter}}
-\item 
-class \mbox{\hyperlink{a00488}{Enum\+Extension\+Methods}}
-\item 
-class \mbox{\hyperlink{a00480}{Enum\+Helper}}
-\item 
-class \mbox{\hyperlink{a00484}{Enum\+To\+Collection\+Converter}}
-\item 
-class \mbox{\hyperlink{a00492}{Formatted\+String\+To\+Nullable\+Double\+Converter}}
-\item 
-class \mbox{\hyperlink{a00496}{Inverse\+Boolean\+Converter}}
-\item 
-class \mbox{\hyperlink{a00500}{Null\+Double\+Helper}}
-\item 
-class \mbox{\hyperlink{a00504}{Sqlite\+Date\+Time\+Helper}}
-\item 
-class \mbox{\hyperlink{a00508}{String\+To\+Nullable\+Double\+Converter}}
-\item 
-class \mbox{\hyperlink{a00512}{String\+To\+Nullable\+U\+Long\+Converter}}
-\item 
-class \mbox{\hyperlink{a00516}{Value\+Description}}
-\end{DoxyCompactItemize}
