+++ Docs/Doxygen/latex/a00415.tex
@@ -1,28 +1,0 @@
-\hypertarget{a00415}{}\doxysection{Audiometry.\+View\+Model Namespace Reference}
-\label{a00415}\index{Audiometry.ViewModel@{Audiometry.ViewModel}}
-\doxysubsection*{Namespaces}
-\begin{DoxyCompactItemize}
-\item 
-namespace \mbox{\hyperlink{a00416}{Bithermal\+Caloric\+VM}}
-\item 
-namespace \mbox{\hyperlink{a00417}{Impedance\+VM}}
-\item 
-namespace \mbox{\hyperlink{a00418}{Patient\+VM}}
-\item 
-namespace \mbox{\hyperlink{a00419}{Pure\+Tone\+VM}}
-\item 
-namespace \mbox{\hyperlink{a00426}{Search\+VM}}
-\item 
-namespace \mbox{\hyperlink{a00427}{Speech\+VM}}
-\end{DoxyCompactItemize}
-\doxysubsection*{Classes}
-\begin{DoxyCompactItemize}
-\item 
-class \mbox{\hyperlink{a00728}{Main\+View\+Model}}
-\item 
-class \mbox{\hyperlink{a00996}{Sub\+Tab\+VM}}
-\item 
-class \mbox{\hyperlink{a01000}{Tab\+VM}}
-\item 
-class \mbox{\hyperlink{a01004}{View\+Model\+Base}}
-\end{DoxyCompactItemize}
