+++ Docs/Doxygen/latex/a00383.tex
@@ -1,16 +1,0 @@
-\hypertarget{a00383}{}\doxysection{Audiometry/\+View\+Model/\+Speech\+V\+M/\+Speech\+Score.cs File Reference}
-\label{a00383}\index{Audiometry/ViewModel/SpeechVM/SpeechScore.cs@{Audiometry/ViewModel/SpeechVM/SpeechScore.cs}}
-\doxysubsection*{Classes}
-\begin{DoxyCompactItemize}
-\item 
-class \mbox{\hyperlink{a00988}{Audiometry.\+View\+Model.\+Speech\+V\+M.\+Speech\+Score}}
-\end{DoxyCompactItemize}
-\doxysubsection*{Namespaces}
-\begin{DoxyCompactItemize}
-\item 
-namespace \mbox{\hyperlink{a00401}{Audiometry}}
-\item 
-namespace \mbox{\hyperlink{a00415}{Audiometry.\+View\+Model}}
-\item 
-namespace \mbox{\hyperlink{a00427}{Audiometry.\+View\+Model.\+Speech\+VM}}
-\end{DoxyCompactItemize}

[…]
