### 2020-01-22 d9c6ad4
Moved the docs folder under audiometry repository.
