+++ docs/Doxygen/latex/a00576.tex
@@ -1,0 +1,236 @@
+\hypertarget{a00576}{}\doxysection{Audiometry.\+Model.\+Sub\+Tab\+Model.\+Stenger\+Model Class Reference}
+\label{a00576}\index{Audiometry.Model.SubTabModel.StengerModel@{Audiometry.Model.SubTabModel.StengerModel}}
+
+
+Inheritance diagram for Audiometry.\+Model.\+Sub\+Tab\+Model.\+Stenger\+Model\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=241pt]{a00575}
+\end{center}
+\end{figure}
+
+
+Collaboration diagram for Audiometry.\+Model.\+Sub\+Tab\+Model.\+Stenger\+Model\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[height=550pt]{a00574}
+\end{center}
+\end{figure}
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\hyperlink{a00576_abee35307ffeb7ff4dc594f2b0405bc37}{Stenger\+Model}} (\mbox{\hyperlink{a00996}{Sub\+Tab\+VM}} vm)
+\item 
+override bool \mbox{\hyperlink{a00576_aa1e784ccbecb9e0b6aec303e5a306074}{Add\+Record\+To\+Database}} (string acc\+Number, string exam\+Date, bool test\+Conducted, out string msg)
+\item 
+override bool \mbox{\hyperlink{a00576_a7b24944833b110c7383889dbf2262927}{Update\+Record\+In\+Database}} (string acc\+Number, string exam\+Date, bool test\+Conducted, out string msg)
+\item 
+override void \mbox{\hyperlink{a00576_a032abf109365a9841828182cacaad3c6}{Open\+Record\+From\+Database}} (string acc\+Number, string exam\+Date, out bool test\+Conducted)
+\item 
+override bool \mbox{\hyperlink{a00576_ac37a922daf609f35d5b63e465812ebb9}{Delete\+Record\+From\+Database}} (string acc\+Number, string exam\+Date)
+\end{DoxyCompactItemize}
+\doxysubsection*{Private Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+void \mbox{\hyperlink{a00576_a5d1242491429bf4751db3fc4938fe39d}{Get\+View\+Model\+Property\+Values}} (string acc\+Number, string exam\+Date, bool test\+Conducted)
+\item 
+string \mbox{\hyperlink{a00576_a52b2b5adf6df0e85e497a468b580065b}{Construct\+Insert\+Query}} ()
+\item 
+string \mbox{\hyperlink{a00576_a4847d63dd755bc029939b1b815376b04}{Construct\+Update\+Query}} ()
+\end{DoxyCompactItemize}
+\doxysubsection*{Static Private Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+static bool \mbox{\hyperlink{a00576_aa1fec1017e12c577662625c10181f88d}{Execute\+Query}} (string query, out string msg)
+\end{DoxyCompactItemize}
+\doxysubsection*{Private Attributes}
+\begin{DoxyCompactItemize}
+\item 
[…]
