+++ docs/Doxygen/latex/a00484.tex
@@ -1,0 +1,57 @@
+\hypertarget{a00484}{}\doxysection{Audiometry.\+Converter.\+Enum\+To\+Collection\+Converter Class Reference}
+\label{a00484}\index{Audiometry.Converter.EnumToCollectionConverter@{Audiometry.Converter.EnumToCollectionConverter}}
+
+
+Inheritance diagram for Audiometry.\+Converter.\+Enum\+To\+Collection\+Converter\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=218pt]{a00483}
+\end{center}
+\end{figure}
+
+
+Collaboration diagram for Audiometry.\+Converter.\+Enum\+To\+Collection\+Converter\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=218pt]{a00482}
+\end{center}
+\end{figure}
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+object \mbox{\hyperlink{a00484_aa8aba32362372e797be2f35e15255692}{Convert}} (object value, Type target\+Type, object parameter, Culture\+Info culture)
+\item 
+object \mbox{\hyperlink{a00484_a7c731dc1b4f73a8520a97097ee664596}{Convert\+Back}} (object value, Type target\+Type, object parameter, Culture\+Info culture)
+\end{DoxyCompactItemize}
+
+
+\doxysubsection{Member Function Documentation}
+\mbox{\Hypertarget{a00484_aa8aba32362372e797be2f35e15255692}\label{a00484_aa8aba32362372e797be2f35e15255692}} 
+\index{Audiometry.Converter.EnumToCollectionConverter@{Audiometry.Converter.EnumToCollectionConverter}!Convert@{Convert}}
+\index{Convert@{Convert}!Audiometry.Converter.EnumToCollectionConverter@{Audiometry.Converter.EnumToCollectionConverter}}
+\doxysubsubsection{\texorpdfstring{Convert()}{Convert()}}
+{\footnotesize\ttfamily object Audiometry.\+Converter.\+Enum\+To\+Collection\+Converter.\+Convert (\begin{DoxyParamCaption}\item[{object}]{value,  }\item[{Type}]{target\+Type,  }\item[{object}]{parameter,  }\item[{Culture\+Info}]{culture }\end{DoxyParamCaption})\hspace{0.3cm}{\ttfamily [inline]}}
+
+Here is the call graph for this function\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=350pt]{a00484_aa8aba32362372e797be2f35e15255692_cgraph}
+\end{center}
+\end{figure}
+\mbox{\Hypertarget{a00484_a7c731dc1b4f73a8520a97097ee664596}\label{a00484_a7c731dc1b4f73a8520a97097ee664596}} 
+\index{Audiometry.Converter.EnumToCollectionConverter@{Audiometry.Converter.EnumToCollectionConverter}!ConvertBack@{ConvertBack}}
+\index{ConvertBack@{ConvertBack}!Audiometry.Converter.EnumToCollectionConverter@{Audiometry.Converter.EnumToCollectionConverter}}
+\doxysubsubsection{\texorpdfstring{ConvertBack()}{ConvertBack()}}
[…]
