### 2020-01-22 e1db148
Corrected the name of Docs folder to docs.
