### 2020-01-23 81f4751
Updated the bug report template
