+++ .github/bug_report.md
@@ -1,6 +1,6 @@
 ---
-name: 🐛 Bug Report
-about: If something isn't working as expected 🤔.
+name: Bug Report
+about: If something isn't working as expected.
 ---
 
 **Describe the bug**
@@ -22,16 +22,7 @@ If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
 
-* OS: [e.g. iOS]
-* Browser [e.g. chrome, safari]
-* Version [e.g. 22]
-
-**Smartphone (please complete the following information):**
-
-* Device: [e.g. iPhone6]
-* OS: [e.g. iOS8.1]
-* Browser [e.g. stock browser, safari]
-* Version [e.g. 22]
+* OS: [e.g. Windows 10]
 
 **Additional context**
 Add any other context about the problem here.
