### 2020-01-23 6140f16
Updated the feature request template
