+++ .github/feature_request.md
@@ -1,6 +1,6 @@
 ---
 name: Feature Request
-about: Request a feature here, or visit discuss.grakn.ai for ideas and questions
+about: Request a feature here
 labels: feature
 ---
 
@@ -12,11 +12,11 @@ Please replace every line in curly brackets { like this } with appropriate answe
 
 ## Current Workaround
 
-{ Please describe how you currently solve or work around this problem, given Grakn's limitation. }
+{ Please describe how you currently solve or work around this problem, given Audiometry's limitation. }
 
 ## Proposed Solution
 
-{ Please describe the solution you would like Grakn to provide, to solve the problem above. }
+{ Please describe the solution you would like Audiometry to provide, to solve the problem above. }
 
 ## Additional Information
 
