### 2020-01-23 8f7da29
Updated the bug report template
