### 2020-02-11 #2016 @martinmodrak
@akeshavan, yes I can look into it, although I've never worked with WPF, so my review will be limited, as I can't really help wit the GUI code (which might be substantial part of the app)
