### 2020-02-19 #2016 @akeshavan
thanks @martinmodrak! Let @drwaseemsheikh know if you have trouble evaluating anything on the checklist (e.g. installation, functionality), since this is useful feedback!
