### 2020-02-21 drwaseemsheikh/audiometry#7 @martinmodrak
I can't see patients I just created
I created and saved two patients (both without using anything except the "Patient" tab). I am unable to find them in the database.
