### 2020-02-21 drwaseemsheikh/audiometry#12 @martinmodrak
I can't find the research-related functionality
The JOSS paper states that:
> 1. The software can be used to compare the sensitivity (true-positive rate) and specificity
> (true-negative rate) of various hearing test equipment and methods. For example,
> questions like how reliable is a pure-tone audiometry test performed by a smartphone or
> a tablet when compared to a benchmark calibrated audiometer can be easily answered
> by using this software.
> 2. The software can be used to determine important correlations between lifestyle, work
> conditions, and demographics; and the types of hearing loss.
> 3. The software can be used to measure the efficacy of a certain treatment, intervention,
> or equipment on the progression of a hearing loss.

I was however unable to locate any functionality related to those tasks neither in code nor in documentation. Where can I find those?
