### 2020-02-21 #2016 @martinmodrak
So I completed a first round of review. I think developing open source software for medical practitioners is a worthwhile endeavour and can be especially valuable for those operating in less developed countries.  The software had some major problems on my computer (e.g. https://github.com/drwaseemsheikh/audiometry/issues/7) , but I believe those are likely very simple bugs that can be easily resolved. I like how the tabs for diagnosis update immediately and they look pretty neat.

There are many small ways the software could be improved to facilitate faster daily use by a practitioner, but I understand if the authors don't have resources to completely polish the interface.

The biggest issue is however that I couldn't find the research-related functionality which is alluded to in the paper anywhere (https://github.com/drwaseemsheikh/audiometry/issues/12), I hope this is just me not looking in the right places. I will resume my review when I hear back from @drwaseemsheikh on this issue.
