### 2020-02-21 drwaseemsheikh/audiometry#12 @drwaseemsheikh
1. You can compare the results of two different hearing test instruments by storing the data from both the instruments and opening their respective results and looking at the changes.

2. You can store the data from individuals from different professions and compare their audiograms or scores and draw conclusions.

3. You can store a progression of tests over time while a patient is under some treatment and look at the plots and data to draw conclusions about the efficacy of the treatment.
