### 2020-02-21 drwaseemsheikh/audiometry#7 @drwaseemsheikh
It doesn't make sense to save a patient record without any test conducted. Can you think of use case for that in a clinical setting? The software is behaving as expected.
