### 2020-02-21 #2016 @drwaseemsheikh
I have addressed all your comments. I would like to reiterate again that as a reviewer your job is to make sure that the software meets the above checkmarks or requirements for the review and not that it is perfect. A lot of changes you asked me to do are small from a user's perspective but would require months of development time from my end which I don't have. I would urge you as a reveiwer to keep into consideration such constraints.

The software has already been downloaded hundreds of times by various users and has received very favorable feedback (See SourceForge). It helps the patients and healthcare providers in underdeveloped countries to better manage and analyze hearing test data thus making the process of hearing impairment diagnosis easier and accurate. Such things also need to be kept in consideration when you are criticizing minor changes.
