### 2020-02-21 #2016 @akeshavan
@drwaseemsheikh, I understand the challenges of developing and maintaining open source software, and that it can be difficult to incorporate reviewers' suggestions when you are limited on time and are working on this software for free. Every author on JOSS experiences this.

But JOSS has standards for the open source software papers that we publish. We are very grateful that our reviewers voluntarily take the time to install our author's software, read their documentation, and ensure functionality and performance, among other things. A reviewer's suggestions, however major or minor, _are not meant to antagonize the author_ -- they are simply suggestions to improve the accessibility and function of the author's software, so that it can reach the broader scientific community.

It is your job as a JOSS author to decide whether to incorporate a reviewer's suggestions or not, keeping in mind the JOSS criteria for publishing. If a reviewer is confused about something, you may want to consider updating your documentation. It is your choice on whether to incorporate any major or minor changes suggested by the reviewer, but I do ask that you respond politely to the reviewer who has volunteered their time to evaluate your software for JOSS.
