### 2020-02-24 #2016 @martinmodrak
@drwaseemsheikh First, I admit that some of the phrasing I used in the issues could be read as combative and I apologize for that. I totally agree with @akeshavan that you don't have to incorporate the feedback I gave into your software. I however do disagree that is my job just to "make sure that the software meets the above checkmarks or requirements", the JOSS review guidelines state that "A key goal of JOSS is to raise the quality of research software generally" and this is the place I am coming from. You may disagree about specific suggestions I made and I am open to have my mind changed about most of them or let @akeshavan as the editor decide that the software can be accepted without resolving them. All of my suggestions were meant to improve the software and I tried to suggest only things I believe would be easy to do (though I might be wrong).

I agree that if there already is a substantial user community, that is a strong plus. If the software does not have GitHub as its primary home but is instead mainly hosted on SourceForge and has a community there, than linking to the SourceForge page in README.md and the paper would IMHO be very sensible.

I will react in more detail at the individual issues.
