### 2020-02-24 drwaseemsheikh/audiometry#12 @martinmodrak
Maybe this is a misunderstanding, but you seem to imply that the software doesn't let me to do this directly - I see how I can store results from different instruments, print the curves and compare them by eye, but it seems that the program doesn't have a "Compare instruments" function that would e.g. overlay "Pure tone" curves from two instruments or compute some distance metrics between the two instruments?

Similarly when I "store the data from individuals from different professions and compare" or follow a patient through treatment, it seems that I need to either do the comparisons manually or open the associated SQLite database in a different program and perform the analysis there. Is that correct?
