### 2020-02-24 drwaseemsheikh/audiometry#7 @martinmodrak
1. The issue persists even when I fill in some values in the "Pure Tone" tab. I am happy to send you some logs or anything else you might need to debug this, it is possible it is related to my computer setup (I am on Windows 10, English edition).
2. I believe there is a use case for documenting only symptoms and diagnosis ("Exam Information" box on the "Patient" tab) without conducting tests or for the case when the examination is for any reason interrupted before conducting a test, in which case I might not want to fill in the patient data again once the examination resumes.
3. If I fill in something, click "Save", the program tells me that "Patient record with account number 1 and date of exam 2/24/2020 12:30:11 PM written to the database." and then nothing is saved, it is definitely a bug. If for some reason the software cannot/should not save patient without any tests conducted, the message upon clicking "Save" should explicitly say something like "No tests conducted, patient data will NOT be saved".
