### 2020-03-10 #2016 @martinmodrak
From my side there are two major issues that block further progress:
- Save/load doesn't work on my computer: https://github.com/drwaseemsheikh/audiometry/issues/7
- The connection to research is unclear https://github.com/drwaseemsheikh/audiometry/issues/12 (this might be actually resolved just by clarifying what the software intends to do and what it does not).

On both I am waiting for reaction from @drwaseemsheikh.
