### 2020-04-21 49c481c
Bug Fixes
Fixed the following bugs:
1. If the symptoms or diagnosis list was empty, the application was not opening the record from the database. It was, however, saving it.
2. I moved the Account textbox to the top and added a tooltip to notify the user that this is a required field.
3. The application was not saving a patient record without any test data.
4. Oxyplot was throwing a null dereference exception when only one value was entered in a textbox for Speech or Impedance audiometry.
