+++ Audiometry/obj/x86/Debug/View/MainWindow.g.i.cs
@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\View\MainWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "D81928F6749C0878A32EAD41AE861C703427E388"
+﻿#pragma checksum "..\..\..\..\View\MainWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "F0E0714AE4513D13FEDD233035A77E7D47B35A29"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -66,7 +66,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 527 "..\..\..\..\View\MainWindow.xaml"
+        #line 526 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.Canvas ExamInfoCanvas;
         
@@ -74,7 +74,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 549 "..\..\..\..\View\MainWindow.xaml"
+        #line 548 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal Xceed.Wpf.Toolkit.DateTimePicker dtExamDateTime;
         
@@ -82,7 +82,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 554 "..\..\..\..\View\MainWindow.xaml"
+        #line 553 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ListBox lbSymptoms;
         
@@ -90,7 +90,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 560 "..\..\..\..\View\MainWindow.xaml"
+        #line 559 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.ListBox lbDiagnosis;
         
@@ -98,7 +98,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 580 "..\..\..\..\View\MainWindow.xaml"
+        #line 579 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.GroupBox PureToneGrpBox;
         
@@ -106,7 +106,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 584 "..\..\..\..\View\MainWindow.xaml"
+        #line 583 "..\..\..\..\View\MainWindow.xaml"
         [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1823:AvoidUnusedPrivateFields")]
         internal System.Windows.Controls.Canvas PureToneAMCanvas;
         
@@ -114,7 +114,7 @@ namespace Audiometry.View {
         #line hidden
         
         
-        #line 586 "..\..\..\..\View\MainWindow.xaml"
[…]
