+++ Audiometry/obj/x86/Debug/View/LoginWindow.g.i.cs
@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\View\LoginWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "10DA92885D954139D54E50973099059124D40B12"
+﻿#pragma checksum "..\..\..\..\View\LoginWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "329F7B04B9347CEDA0F32883BE80E0611EB70F54"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
