+++ Audiometry/obj/x86/Debug/View/SearchWindow.g.i.cs
@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\View\SearchWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "A5C9B27131B6A3BE21880A948A02FE0FB9EFB862"
+﻿#pragma checksum "..\..\..\..\View\SearchWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "8796303A6641B4925B65EC849F031887E077D5B8"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
