+++ AudiometryWixInstaller/obj/Debug/AudiometryWixInstaller.wixproj.BindBuiltOutputsFileListnull.txt
@@ -1,2 +1,0 @@
-E:\SoftwareProjects\AudiometryApplication\trunk\src\Audiometry\AudiometryWixInstaller\bin\Debug\AudiometryWixInstaller.msi
-E:\SoftwareProjects\AudiometryApplication\trunk\src\Audiometry\AudiometryWixInstaller\bin\Debug\AudiometryWixInstaller.wixpdb
