### 2020-04-21 #2016 @drwaseemsheikh
I have fixed the following bugs which include the most important issues that were raised during the review process. I have committed my code changes and updated the installer binary file. You can test and verify that these issues have been fixed by installing the installer file under the install folder.

Fixed the following bugs:
1. If the symptoms or diagnosis list was empty, the application was not opening the record from the database. It was, however, saving it.
2. I moved the Account textbox to the top and added a tooltip to notify the user that this is a required field.
3. The application was not saving a patient record without any test data.
4. Oxyplot was throwing a null dereference exception when only one value was entered in a textbox for Speech or Impedance audiometry.

As per the changes that you want with respect to the research applications, why don't you tell me how would you like the wording? I will be happy to change the wording in the paper.
