### 2020-04-21 1d28efa
Fixed SISI and ToneDecay textbox binding issue
These textboxes were not recording decimal values correctly.
