+++ Audiometry/obj/x86/Debug/View/MainWindow.g.i.cs
@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\View\MainWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "F0E0714AE4513D13FEDD233035A77E7D47B35A29"
+﻿#pragma checksum "..\..\..\..\View\MainWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "08BAD33FBB5465BE4BFF552534BB695C2F117070"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
