### 2020-05-05 #2016 @arfon
Hi @waseemasheikh / @drwaseemsheikh - it's not clear if you have addressed the feedback in drwaseemsheikh/audiometry#7 or drwaseemsheikh/audiometry#12. Could you please comment on that?

> Could you please let me know the timeline for review and publishing as the paper has been under review for a very long time.

As I noted in [my comment](https://github.com/openjournals/joss-reviews/issues/2016#issuecomment-598981503) in mid-March, JOSS is working on a best-efforts basis right now. The effects of the pandemic on the availability of our reviewers and editors to contribute time to reviews is hard to gauge. Your patience is appreciated.
