### 2020-05-12 #2016 @drwaseemsheikh
Hi.

Regarding Issue #7 (where a patient's record was not being saved when only the patient tab was filled in): This bug has been fixed in my last commit.

Regarding Issue #12 (research applications): The co-author of this paper (Dr. Nadeem) is a board-certified ENT specialist and he is using this software in his research in the following specific contexts.

1. The software can be used to compare the sensitivity (true-positive rate) and specificity (true-negative rate) of various hearing test equipment and methods. For example, questions like how reliable is a pure-tone audiometry test performed by a smartphone or a tablet when compared to a benchmark calibrated audiometer can be easily answered by using this software.

How: Record the audiometry data from the two different tests, generated the audiogram plots for the two different sets of data, and compare them to look for differences in thresholds.

2. The software can be used to determine important correlations between lifestyle, work conditions, and demographics; and the types of hearing loss.

How: The co-author used this software in collecting hearing test data for airmen who were exposed to loud noise. Their audiograms were generated and the ENT specialist was able to see certain trends in their audiograms which predicted a certain kind of hearing loss.

3. The software can be used to measure the efficacy of a certain treatment, intervention, or equipment on the progression of a hearing loss.

How: Record the hearing test data of a patient during multiple stages of an intervention. Compare the audiograms and observe the trends.
