### 2020-05-13 #2016 @martinmodrak
Sorry for the delay. I've just installed an updated version (it is a bit hard to verify that this is the latest version, as the version in the Audiometry.exe metadata has not changed and is still 1.0.0, but the file date is April 20th, so I guess this is correct).

Issue #7 persists, I am still unable to save patients and their data.

As for the scientific use: I think saying that the fact that you can record and retrieve medical information and then compare it without any help from the program stretches the definition of "obvious research application" way beyond what I understand by those words (a text editor allows you to do this as well). Since this is my first time reviewing, I however don't have a strong sense on how strongly this is usually interpreted and will delegate the decision on this to @arfon.
