### 2020-05-13 drwaseemsheikh/audiometry#7 @martinmodrak
The issue persists after updating to latest version, even if I fill some values in the "Pure tone" tab.
