### 2020-05-25 drwaseemsheikh/audiometry#7 @drwaseemsheikh
I am unable to reproduce this bug. I will request Afron to ask another reviewer to test this issue and provide detailed steps of how the bug is introduced. Just saying that the issue persists which I can't replicate is not helpful. I am unable to reproduce this bug and when I save the patient data alone, I am able to save and restore it. So we are at an impasse and I would ask the editor to find a way out of this.
