### 2020-05-25 #2016 @drwaseemsheikh
I am unable to replicate issue # 7 bug. I have tested it on my end and that issue has been fixed. I am able to save and restore patient data without any data. Since we are at an impasse now, I would ask the editor or another reviewer to test the functionality of the code and provide detailed steps so that I can replicate that issue. Just saying that the issue persists doesn't help me as I can't replicate it. To the best of my understanding and testing, that issue has been fixed.
