### 2020-05-25 #2016 @drwaseemsheikh
Here is the proof with the screeshots that bug#7 has been fixed. Steps to follow:
1. Install the latest application.
2. Enter data in the Patient tab, leaving all other tabs blank.
3. Save the patient data by clicking on Save.
4. Click on the Search menu option and set the start date in the Search window to the previous day or any time before the data was recorded. Click Search.
5. This shows the patient record that was just saved. Opening this record shows the exact data filled in the form as was saved. See pictures for the proof.

My guess is (since the reviewer has not provided detailed steps) that he is not setting the Start date/time field in the Search window to before the date/time of the test taken and therefore he doesn't see that record. But that's an error on his part and not on the software.
![fig1](https://user-images.githubusercontent.com/56233110/82842794-24552980-9e98-11ea-8642-e83ebe97a680.png)
![fig2](https://user-images.githubusercontent.com/56233110/82842795-24edc000-9e98-11ea-9a89-334bd548a397.png)
![fig3](https://user-images.githubusercontent.com/56233110/82842796-24edc000-9e98-11ea-9f9d-583b64396c32.png)
![fig4](https://user-images.githubusercontent.com/56233110/82842797-24edc000-9e98-11ea-8a9d-428bb04be6cb.png)
![fig5](https://user-images.githubusercontent.com/56233110/82842798-25865680-9e98-11ea-96ff-894e4bce978e.png)
