### 2020-05-25 #2016 @arfon
> Thanks for the update @drwaseemsheikh. Regarding the usage in research, are there any papers/publications that could be cited?

@drwaseemsheikh - could you please respond to this question I posed a little while back?
