### 2020-05-26 #2016 @martinmodrak
@drwaseemsheikh I just verified one more time the problem https://github.com/drwaseemsheikh/audiometry/issues/7 persists on my work PC. I will also try another machine with a different system. However, the fact that software behaves differently on different computers is unfortunately to be expected and it is generally considered good practice to test on multiple computers and resolve problems that appear. I don't think it is my role to debug this issue in depth - I tried some obvious steps (e.g. running the software as admin) but those didn't help. But as I said in the issue, I am happy to provide any information that you think might be related (e.g. logs from your application) or run some test application/modified version that provides extensive info that would help you debug. If you prefer, we can possibly also arrange a call with screen sharing. I really have no intention to shoot the application down and I want to help to make the app better. If it doesn't work on my system, it likely fails on systems of many other potential users, but most users don't report bugs and instead just stop using the software.
