### 2020-05-26 #2016 @arfon
> @drwaseemsheikh I just verified one more time the problem drwaseemsheikh/audiometry#7 persists on my work PC. I will also try another machine with a different system. However, the fact that software behaves differently on different computers is unfortunately to be expected and it is generally considered good practice to test on multiple computers and resolve problems that appear.

@martinmodrak - thanks. I think that would be very helpful. I agree with the spirit of your comment here and think it would be good to, if at all possible, resolve this issue before closing out this review.
