### 2020-05-27 drwaseemsheikh/audiometry#7 @martinmodrak
I tried on a different PC (Windows 7, .NET Framework 4.7.02053) and I still can't save the patient. I followed exactly the steps you mentioned at https://github.com/openjournals/joss-reviews/issues/2016#issuecomment-633714396 . The confirmation dialog (patient saved) appears, but when I click "Open" the list of patients is empty. I would suggest you also try to install and run Audiometry on a few different PCs and see if any of them has the same issue. I reiterate that if you want any diagnostic info, I'll be happy to provide it.
