### 2020-05-28 drwaseemsheikh/audiometry#7 @drwaseemsheikh
Bug can't be reproduced after installing the software on several machines. This issue has been fixed.
