### 2020-05-28 drwaseemsheikh/audiometry#12 @drwaseemsheikh
The research related functionality has been explained in the threads numerous times.
