### 2020-05-29 #2016 @martinmodrak
Looking at the progress of the individual issues I filed, it is clear I was unable to build a collaborative relationship with the author.  I have no aim to gatekeep the publication of the software, but I am also unable to verify some of the basic functionality and can't find meaningful way forward. This is a lose-lose situation and I therefore ask @arfon to remove me from reviewing this submission.
