### 2020-05-29 #2016 @arfon
Thank you @martinmodrak for all of your hard work thus far. I'm sorry things have reached this point but I understand why you would like to be removed at this stage.

@drwaseemsheikh - while I look for a new reviewer, could you please address my comment from a couple of weeks ago about [published research](https://github.com/openjournals/joss-reviews/issues/2016#issuecomment-627436937) that you might be able to cite here?
