### 2020-05-29 #2016 @arfon
@drwaseemsheikh - I've just tried installing this software on a Windows 10 Pro VM on Azure and confirmed that I can complete the [steps described by the author](https://github.com/openjournals/joss-reviews/issues/2016#issuecomment-633714396) to save a record.

As such, I believe we can move forward with accepting this submission.

@martinmodrak - given all of the hard work you've put into this review, I'd rather leave you listed as a reviewer here if that is acceptable to you? (That way you'll be listed as a reviewer on the final paper). Please let me know if this is not OK.

@drwaseemsheikh - if there are any published papers/articles that have made use of this software, please cite them. Otherwise:

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
