### 2020-07-01 #2016 @drwaseemsheikh
Here are the updates from my end:

1. There are no published papers/articles that have made use of this software at this point.

2. The software release is the latest and has all the bug fixes and changes from past reviews.

3. I have archived the software in Zenodo and following is the DOI of the archive

DOI: 10.5281/zenodo.3924784

The title of the archive is the same as the JOSS paper title and the authors of the archive are the same as the JOSS paper authors.
