### 2020-07-01 #2016 @arfon
@drwaseemsheikh - could you review this list of DOIs ☝️ for your references and for the ones that look correct, please add them to your BibTeX file? For example if `https://doi.org/10.1055/b-006-161125 may be missing for title: Essentials of Audiology` is correct for this title, then add `doi=10.1055/b-006-161125` to the BibTeX entry for that paper.

@openjournals/joss-eics - apologies - I forgot to check the references here before running the proof-generation stage. This submission isn't quite ready for processing.
