### 2020-07-01 afdfc80
Updated the bibtex file with missing DOIs
