+++ paper/paper.bib
@@ -44,6 +44,7 @@
   author={Gelfand, Stanley A},
   year={2016},
   edition={4},
+  doi={10.1055/b-006-161125},
   publisher={Thieme Medical Publishers}
 }
 
@@ -52,6 +53,7 @@
   author={Katz, Jack and Chasin, Marshall and English, Kristina M and Hood, Linda J and Tillery, Kim L},
   year={2015},
   edition={7},
+  doi={10.1097/00003446-198608000-00010},
   publisher={Wolters Kluwer Health}
 }
 
@@ -104,6 +106,7 @@
   booktitle={2017 IEEE 3rd Information Technology and Mechatronics Engineering Conference (ITOEC)},
   pages={466--469},
   year={2017},
+  doi={10.1109/itoec.2017.8122338},
   organization={IEEE}
 }
 
@@ -115,6 +118,7 @@
   number={1},
   pages={170--179},
   year={2018},
+  doi={10.1109/tbcas.2018.2878341},
   publisher={IEEE}
 }
 
@@ -149,6 +153,7 @@
   number={9},
   pages={697--704},
   year={2015},
+  doi={10.1089/tmj.2014.0171},
   publisher={Mary Ann Liebert, Inc. 140 Huguenot Street, 3rd Floor New Rochelle, NY 10801 USA}
 }
 
@@ -171,6 +176,7 @@
   number={5},
   pages={700--706},
   year={2013},
+  doi={10.1177/0194599813501461},
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
 }
 
@@ -182,6 +188,7 @@
   number={7},
   pages={678--681},
   year={2009},
+  doi={10.1089/tmj.2009.0013},
   publisher={Mary Ann Liebert, Inc. 140 Huguenot Street 3rd Floor New Rochelle, NY 10801 USA}
 }
 
@@ -193,6 +200,7 @@
   number={4},
   pages={375--378},
   year={2003},
+  doi={10.1089/153056203772744707},
   publisher={Mary Ann Liebert, Inc.}
 }
 
@@ -204,6 +212,7 @@
   number={1},
   pages={95--103},
   year={1996},
+  doi={10.1016/0169-2607(95)01712-7},
   publisher={Elsevier}
 }
 
@@ -213,6 +222,7 @@
   journal={Journal of telemedicine and telecare},
   pages={1357633X18800856},
   year={2018},
+  doi={10.1177/1357633x18800856},
   publisher={SAGE Publications Sage UK: London, England}
 }
 
@@ -224,6 +234,7 @@
   number={9},
   pages={747--752},
   year={2017},
+  doi={10.1089/tmj.2016.0253},
   publisher={Mary Ann Liebert, Inc. 140 Huguenot Street, 3rd Floor New Rochelle, NY 10801 USA}
 }
 
@@ -235,6 +246,7 @@
   number={4},
   pages={570--580},
   year={2018},
+  doi={10.1044/2018_aja-17-0070},
   publisher={ASHA}
 }
 
@@ -246,6 +258,7 @@
   number={2},
   pages={333--339},
   year={2016},
+  doi={10.1007/s00405-015-3533-9},
   publisher={Springer}
 }
 
@@ -267,7 +280,8 @@
   volume={12},
   number={1},
   pages={61--6},
-  year={2016}
+  year={2016},
+  doi={10.5152/iao.2016.1421}
 }
 
 @article{HandzelSmartphone:2013,
@@ -278,6 +292,7 @@
   number={4},
   pages={201--207},
   year={2013},
+  doi={10.1159/000349913},
   publisher={Karger Publishers}
 }
 
@@ -289,6 +304,7 @@
   number={3},
[…]
