### 2020-07-01 #2016 @drwaseemsheikh
I have updated my BibTex file with the missing DOIs.
